refactor: Chat Message Assistant componentization
